+++ src/api/catalog-page.js
@@ -11,4 +11,11 @@ function getCatalogPageData({ limit, offset }) {
     .catch((err) => Promise.reject(new Error(`${err.message}`)));
 }
 
-export default getCatalogPageData;
+function getCatalogArticlePageData({ articleId } = {}) {
+  return axios
+    .get(`${baseURL}${apiUrl}/catalog/${articleId}`)
+    .then((response) => response.data)
+    .catch((err) => Promise.reject(err?.response));
+}
+
+export { getCatalogPageData, getCatalogArticlePageData };
